[refactor] OrderItemCommandService.java, DataPlatform.java 패키지 변경
